Fix memory usage when reading input
Currently, every time mod_zip reads 8KB, it allocates a new string, copies all existing data,
and doesn't free the old string. This results in GBs of memory use for a 10MB input.

Just use an nginx array which automatically doubles in size as needed.
(Still leaks memory, but not as much.)